% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cr_graphs.R
\name{generate_cr_graphs}
\alias{generate_cr_graphs}
\title{Generates serum creatinine graphs for patients.}
\usage{
generate_cr_graphs(
  siteid,
  base_table,
  obfuscation,
  obfuscation_level,
  kdigo_grade_table,
  ckd_valid,
  ckd_table,
  labs_cr,
  use_index_baseline = FALSE,
  index_admit_baselines,
  use_custom_output = FALSE,
  use_custom_output_dir = "/4ceData/Output"
)
}
\arguments{
\item{base_table}{peak_trend}

\item{obfuscation}{is_obfuscated}

\item{obfuscation_level}{obfuscation_value}

\item{kdigo_grade_table}{kdigo_grade}

\item{ckd_valid}{ckd_present}

\item{ckd_table}{ckd_list}

\item{labs_cr}{labs_cr_all}

\item{use_index_baseline}{Indicates if the new definition of baseline sCr (-365 days to last day of index admission) is to be used.}

\item{index_admit_baselines}{baseline_cr_index_admit}
}
\description{
Generates serum creatinine graphs for patients.
}
