% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cr_graphs_prioronly.R
\name{generate_cr_graphs_prioronly}
\alias{generate_cr_graphs_prioronly}
\title{Generates serum creatinine graphs for patients with pre-admission serum creatinine values only.}
\usage{
generate_cr_graphs_prioronly(
  siteid,
  base_table,
  obfuscation,
  obfuscation_level,
  kdigo_grade_table,
  ckd_valid,
  ckd_table,
  labs_cr
)
}
\arguments{
\item{base_table}{peak_trend}

\item{obfuscation}{is_obfuscated}

\item{obfuscation_level}{obfuscation_value}

\item{kdigo_grade_table}{kdigo_grade}

\item{ckd_valid}{ckd_present}

\item{ckd_table}{ckd_list}

\item{labs_cr}{labs_cr_all}
}
\description{
Generates serum creatinine graphs for patients with pre-admission serum creatinine values only.
}
