% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_demog_files.R
\name{generate_demog_files}
\alias{generate_demog_files}
\title{Creates (non)-obfuscated demographic tables and also pre-processes demographics for use in time-to-event analyses.
If specified, can also filter the output to pre-specified patients with pre-admission sCr values.}
\usage{
generate_demog_files(
  siteid,
  demog_table,
  aki_labs,
  comorbid_table,
  comorbid_header,
  kdigo_grade_table,
  meld_valid = FALSE,
  labs_meld_admit = NULL,
  coaga_valid,
  coagb_valid,
  covid_valid,
  remdesivir_valid,
  acei_valid,
  arb_valid,
  med_coaga = NULL,
  med_coagb = NULL,
  med_covid19 = NULL,
  med_acearb = NULL,
  cirrhosis_valid,
  preadmit_cr_list = NULL,
  preadmit_only_analysis = FALSE,
  obfuscation,
  obfuscation_level,
  aki_first_visit_list
)
}
\arguments{
\item{siteid}{currSiteId}

\item{demog_table}{The demographics_filt equivalent}

\item{comorbid_table}{The comorbid equivalent}

\item{comorbid_header}{The comorbid_list equivalent}

\item{kdigo_grade_table}{kdigo_grade}

\item{meld_valid}{meld_analysis_valid}

\item{labs_meld_admit}{labs_meld_admission}

\item{coaga_valid}{coaga_present}

\item{coagb_valid}{coagb_present}

\item{covid_valid}{covid19antiviral_present (Novel COVID-19 antiviral use)}

\item{remdesivir_valid}{remdesivir_present (Remdesivir use)}

\item{acei_valid}{acei_present (ACE-inhibitor use)}

\item{arb_valid}{arb_present (ARB use)}

\item{med_coaga}{med_coaga_new (List of patients with COAGA)}

\item{med_coagb}{med_coagb_new (List of patients with COAGB)}

\item{med_covid19}{med_covid19_new (List of patients with Novel COVID-19 antiviral/remdesivir use)}

\item{med_acearb}{med_acearb_chronic (List of patients previously on ACE-inhibitors/ARBs)}

\item{cirrhosis_valid}{cirrhosis_present (Whether cirrhotic patients are in the cohort)}

\item{preadmit_cr_list}{Pre-specified patient list with pre-admission sCr labs. Default value is NULL}

\item{preadmit_only_analysis}{Specifies whether demographics are to be narrowed down to list specified by preadmit_cr_list. Default is FALSE.}

\item{obfuscation}{is_obfuscated}

\item{obfuscation_level}{obfuscation_value}
}
\value{
A list: demog_toe gives a table with the filtered demographics ready for merging for time-to-event
analyses, while demog_var_list provides the headers of relevant demographics to be included during the automated filtering
process for model construction.
}
\description{
Creates (non)-obfuscated demographic tables and also pre-processes demographics for use in time-to-event analyses.
If specified, can also filter the output to pre-specified patients with pre-admission sCr values.
}
