% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_life_table_competing_risk.R
\name{generate_life_table_competing_risk}
\alias{generate_life_table_competing_risk}
\title{Generates risk tables for combining into a single aggregated life table for computing cumulative incidence curves for AKI recovery and New Onset CKD (competing risk is mortality)}
\usage{
generate_life_table_competing_risk(
  aki_index_recovery,
  aki_index_nonckd,
  aki_index_nonckd_akionly,
  aki_index_ckdonly_akionly,
  use_custom_output = FALSE,
  use_custom_output_dir = "/4ceData/Output"
)
}
\arguments{
\item{aki_index_recovery}{The aki_index_recovery table generated from run_time_to_event_analysis()}

\item{aki_index_nonckd}{The aki_index_nonckd table generated from run_time_to_event_analysis_nonckd()}

\item{aki_index_nonckd_akionly}{The aki_index_nonckd_akionly table generated from run_time_to_event_analysis_nonckd()}

\item{aki_index_ckdonly_akionly}{The aki_index_ckdonly_akionly table generated from run_time_to_event_analysis_nonckd()}
}
\description{
Generates risk tables for combining into a single aggregated life table for computing cumulative incidence curves for AKI recovery and New Onset CKD (competing risk is mortality)
}
