% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrivateSummaryRepositoryUrl.R
\name{getPrivateSummaryRepositoryUrl}
\alias{getPrivateSummaryRepositoryUrl}
\title{Returns the GitHub URL of the Private Summary Repository for this project for the AKI project}
\usage{
getPrivateSummaryRepositoryUrl()
}
\description{
Returns the GitHub URL of the Private Summary Repository for this project for the AKI project
}
\keyword{4CE}
