% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProjectOutputDirectory.R
\name{getProjectOutputDirectory}
\alias{getProjectOutputDirectory}
\title{Returns the name of the project's output directory. runAnalysis() should save its output here, and submitAnalysis() will expect to read from this location. This function checks to make sure the location exists, and if not, creates it. for the AKI project}
\usage{
getProjectOutputDirectory()
}
\description{
Returns the name of the project's output directory. runAnalysis() should save its output here, and submitAnalysis() will expect to read from this location. This function checks to make sure the location exists, and if not, creates it. for the AKI project
}
\keyword{4CE}
