% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_recovery_analysis_150.R
\name{run_recovery_analysis_150}
\alias{run_recovery_analysis_150}
\title{Runs time-to-event analysis for the 4 pre-defined models - Kaplan-Meier curves, Cox PH models.
This version uses the 150\% baseline sCr as the definition for renal recovery instead, and only runs recovery analysis.
If specified, can also filter the output to pre-specified patients with pre-admission sCr values.}
\usage{
run_recovery_analysis_150(
  siteid,
  base_table,
  aki_episodes,
  aki_labs,
  demog_table,
  demog_toe = NULL,
  demog_var_list = NULL,
  comorbid_table = NULL,
  comorbid_header = NULL,
  kdigo_grade_table = NULL,
  ckd_valid,
  coaga_valid,
  coagb_valid,
  covid_valid,
  remdesivir_valid,
  acei_valid,
  arb_valid,
  med_coaga = NULL,
  med_coagb = NULL,
  med_covid19 = NULL,
  med_acearb = NULL,
  preadmit_cr_list = NULL,
  preadmit_only_analysis = FALSE,
  obfuscation,
  obfuscation_level,
  restrict_model_corr,
  factor_threshold = 5
)
}
\arguments{
\item{siteid}{currSiteId (site ID of the site as obtained earlier)}

\item{base_table}{peak_trend}

\item{aki_episodes}{aki_index}

\item{demog_table}{The demographics_filt equivalent}

\item{demog_toe}{Specify the appropriate demog_time_to_event table}

\item{demog_var_list}{Specify the appropriate demog_list list}

\item{comorbid_table}{The comorbid equivalent}

\item{comorbid_header}{The comorbid_list equivalent}

\item{kdigo_grade_table}{kdigo_grade}

\item{ckd_valid}{ckd_present (Whether at least one patient has CKD as a comorbidity)}

\item{coaga_valid}{coaga_present}

\item{coagb_valid}{coagb_present}

\item{covid_valid}{covid19antiviral_present (Novel COVID-19 antiviral use)}

\item{remdesivir_valid}{remdesivir_present (Remdesivir use)}

\item{acei_valid}{acei_present (ACE-inhibitor use)}

\item{arb_valid}{arb_present (ARB use)}

\item{med_coaga}{med_coaga_new (List of patients with COAGA)}

\item{med_coagb}{med_coagb_new (List of patients with COAGB)}

\item{med_covid19}{med_covid19_new (List of patients with Novel COVID-19 antiviral/remdesivir use)}

\item{med_acearb}{med_acearb_chronic (List of patients previously on ACE-inhibitors/ARBs)}

\item{preadmit_cr_list}{Pre-specified patient list with pre-admission sCr labs. Default value is NULL}

\item{preadmit_only_analysis}{Specifies whether demographics are to be narrowed down to list specified by preadmit_cr_list. Default is FALSE.}

\item{obfuscation}{is_obfuscated}

\item{obfuscation_level}{obfuscation_value}
}
\description{
Runs time-to-event analysis for the 4 pre-defined models - Kaplan-Meier curves, Cox PH models.
This version uses the 150\% baseline sCr as the definition for renal recovery instead, and only runs recovery analysis.
If specified, can also filter the output to pre-specified patients with pre-admission sCr values.
}
