% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_time_to_event_analysis_nonckd.R
\name{run_time_to_event_analysis_nonckd}
\alias{run_time_to_event_analysis_nonckd}
\title{Runs time-to-event analysis for the subgroup of non-CKD patients, including time to new onset CKD}
\usage{
run_time_to_event_analysis_nonckd(
  siteid,
  aki_index_recovery,
  aki_index_death,
  med_recovery_list,
  comorbid_recovery_list,
  demog_recovery_list,
  earliest_cr_recovery_list,
  med_death_list,
  comorbid_death_list,
  demog_death_list,
  earliest_cr_death_list,
  obfuscation,
  obfuscation_level,
  restrict_model_corr,
  factor_threshold = 5,
  use_custom_output = FALSE,
  use_custom_output_dir = "/4ceData/Output",
  ckd_present = TRUE,
  input = "/4ceData/Input"
)
}
\arguments{
\item{siteid}{currSiteId (site ID of the site as obtained earlier)}

\item{aki_index_recovery}{aki_index_recovery table from running the run_time_to_event_analysis() function}

\item{aki_index_death}{aki_index_death table from running the run_time_to_event_analysis() function}

\item{obfuscation}{is_obfuscated}

\item{obfuscation_level}{obfuscation_value}

\item{restrict_model_corr}{Whether to restrict models}

\item{factor_threshold}{Threshold to remove a covariate from analysis if the total number of events in a level is less than this number. Default = 5}

\item{use_custom_output}{Specifies if a custom output directory is to be used}

\item{use_custom_output_dir}{Custom output directory name}

\item{lists}{Respective lists from running the run_time_to_event_analysis() function}
}
\description{
Runs time-to-event analysis for the subgroup of non-CKD patients, including time to new onset CKD
}
