% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAnalysis.R
\name{runAnalysis}
\alias{runAnalysis}
\title{Runs the analytic workflow for the AKI project}
\usage{
runAnalysis(is_obfuscated = TRUE, obfuscation_value = 3)
}
\description{
Runs the analytic workflow for the AKI project
}
\keyword{4CE}
