% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitAnalysis.R
\name{submitAnalysis}
\alias{submitAnalysis}
\title{Submits the results of the analytic workflow for the AKI project}
\usage{
submitAnalysis()
}
\description{
Submits the results of the analytic workflow for the AKI project
}
\keyword{4CE}
