% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thromb_icd10_ref.R
\docType{data}
\name{thromb_icd10_ref}
\alias{thromb_icd10_ref}
\title{Diagnoses Codes for Thrombotic Events (ICD10)}
\format{
An object of class \code{data.frame} with 144 rows and 3 columns.
}
\usage{
data(thromb_icd10_ref)
}
\description{
This table contains ICD10 codes for thrombo-embolic events (e.g. pulmonary embolism, myocardial
infarction, deep vein thromboembolism, arterial thromboembolism).
This table is an abbreviated version derived from the main table thromb_ref
}
\details{
The headers in the table are as follows:
icd_code  full_code type

icd_code - shortened ICD code (see below)
full_code - full ICD code (see below)
type - diagnosis category as above

Due to the variations in the way ICD codes may be stored, we have adhered to the following conventions:
\itemize{
\item icd_code will store the broad category of the diagnosis
\item Full ICD diagnosis codes will have the major category abd subcategory separated by a period
For example, the ICD-10 diagnosis "Septic pulmonary embolism with acute cor pulmonale" with full code I260.1
will have icd_code = "I260" and full_code = "I260.1".
Please ensure that your data tables store diagnoses codes as their broad categories. (e.g. I260.1 will be
stored as "I260").
In future versions, this convention may be changed depending on the granularity required for analysis.
}
}
\keyword{datasets}
