% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateAnalysis.R
\name{validateAnalysis}
\alias{validateAnalysis}
\title{Validates the results of the analytic workflow for the AKI project}
\usage{
validateAnalysis(
  docker = TRUE,
  siteid_nodocker = "",
  custom_output = FALSE,
  custom_output_dir = ""
)
}
\description{
Validates the results of the analytic workflow for the AKI project
}
\keyword{4CE}
