% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAnalysis.R
\name{runAnalysis}
\alias{runAnalysis}
\title{Runs the analytic workflow for the AKI project}
\usage{
runAnalysis(
  is_obfuscated = TRUE,
  factor_cutoff = 5,
  ckd_cutoff = 2.25,
  restrict_models = FALSE,
  docker = TRUE,
  input = "/4ceData/Input",
  siteid_nodocker = "",
  skip_qc = FALSE
)
}
\description{
Runs the analytic workflow for the AKI project
}
\keyword{4CE}
